requirements.txt
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.0.0
